requirements.txt
pycryptodome
matplotlib
tqdm
numpy 